requirements.txt
torch>=2.1.0
torchvision
torchaudio
numpy==1.26.4
diffusers==0.29.2
transformers==4.40.2
accelerate
fastapi
uvicorn
pillow
opencv-python
pydantic
python-multipart
plyfile